+++ mex/common/settings.py
@@ -1,16 +1,22 @@
+from os import environ
 from pathlib import Path
+from textwrap import wrap
 from typing import Any, Self, cast
 
+import click
 from pydantic import BaseModel as PydanticBaseModel
 from pydantic import Field, HttpUrl, SecretStr, model_validator
 from pydantic_settings import BaseSettings as PydanticBaseSettings
 from pydantic_settings import SettingsConfigDict
 from pydantic_settings.sources import ENV_FILE_SENTINEL, DotenvType, EnvSettingsSource
+from tabulate import tabulate
 
 from mex.common.context import SingleSingletonStore
+from mex.common.logging import logger
 from mex.common.types import AssetsPath, IdentityProvider, Sink, WorkPath
 
 SETTINGS_STORE = SingleSingletonStore["BaseSettings"]()
+MEX_ASSETS_DIR = "MEX_ASSETS_DIR"
 
 
 class BaseSettings(PydanticBaseSettings):
@@ -48,22 +54,15 @@ class BaseSettings(PydanticBaseSettings):
         _secrets_dir: str | Path | None = None,
         **values: Any,  # noqa: ANN401
     ) -> None:
-        """Construct a new settings instance.
-
-        After building from regular settings sources, we look for a configured
-        `assets_dir` and then check that folder for a dotenv file as well.
-        Because of this dependency from one setting source to another, sadly
-        we cannot use `Config.customise_sources`.
-        """
-        settings_wo_assets_env_file = self._settings_build_values(
-            values,
-            _env_file=_env_file,
-            _env_file_encoding=_env_file_encoding,
-            _env_nested_delimiter=_env_nested_delimiter,
-            _secrets_dir=_secrets_dir,
-        )
-        if assets_dir := settings_wo_assets_env_file.get("MEX_ASSETS_DIR"):
-            _env_file = Path(assets_dir, ".env")
+        """Construct a new settings instance."""
+        if assets_dir := environ.get(MEX_ASSETS_DIR):
+            if not _env_file or _env_file is ENV_FILE_SENTINEL:
+                _env_file = [
+                    Path(assets_dir, ".env"),  # bw-compat
+                    Path(assets_dir, "config", ".env"),
+                ]
+            if not _secrets_dir:
+                _secrets_dir = Path(assets_dir, "config", "secrets")
         super().__init__(
             _env_file=_env_file,
             _env_file_encoding=_env_file_encoding,
@@ -86,8 +85,10 @@ class BaseSettings(PydanticBaseSettings):
 
     debug: bool = Field(
         False,  # noqa: FBT003
-        alias="pdb",
-        description="Jump into post-mortem debugging after any uncaught exception.",
+        description=(
+            "Enable application specific debugging behavior, like expanded logging or "
+            "testing-only functionality."
+        ),
         validation_alias="MEX_DEBUG",
     )
     sink: list[Sink] = Field(
@@ -104,7 +105,7 @@ class BaseSettings(PydanticBaseSettings):
             "Path to directory that contains input files treated as read-only, "
             "looks for a folder named `assets` in the current directory by default."
         ),
-        validation_alias="MEX_ASSETS_DIR",
+        validation_alias=MEX_ASSETS_DIR,
     )
     work_dir: Path = Field(
         Path.cwd(),
@@ -201,17 +202,21 @@ class BaseSettings(PydanticBaseSettings):
         """Dump the current settings into a readable table.
 
         Returns:
-            Formatted string with all settings key-value pairs.
+            Formatted table with all settings.
         """
-        dict_ = self.model_dump()
-        indent = max(len(key) for key in dict_)
-        return "\n".join(
-            [
-                f"{key.ljust(indent)} "
-                f"{', '.join(str(v) for v in val) if isinstance(val, list) else val}"
-                for key, val in dict_.items()
-            ]
-        )
+
+        def format_value(v: object) -> str:
+            return "\n".join(
+                wrap(", ".join(map(str, v)) if isinstance(v, list) else str(v))
+            )
+
+        rows: list[tuple[str, str]] = []
+        for k, v in self.model_dump(mode="json", exclude_unset=not self.debug).items():
+            if isinstance(v, dict):
+                rows.extend((f"{k}__{sk}", format_value(sv)) for sk, sv in v.items())
+            else:
+                rows.append((k, format_value(v)))
+        return tabulate(rows, tablefmt="simple_grid")
 
     @classmethod
     def get_env_name(cls, name: str) -> str:
@@ -243,9 +248,9 @@ class BaseSettings(PydanticBaseSettings):
         def _resolve(model: PydanticBaseModel, _name: str) -> None:
             value = getattr(model, _name)
             if isinstance(value, AssetsPath) and value.is_relative():
-                setattr(model, _name, self.assets_dir.resolve() / value)
+                object.__setattr__(model, _name, self.assets_dir.resolve() / value)
             elif isinstance(value, WorkPath) and value.is_relative():
-                setattr(model, _name, self.work_dir.resolve() / value)
+                object.__setattr__(model, _name, self.work_dir.resolve() / value)
             elif isinstance(value, PydanticBaseModel):
                 for sub_model_field_name in type(value).model_fields:
                     _resolve(value, sub_model_field_name)
@@ -274,4 +279,15 @@ class BaseSettings(PydanticBaseSettings):
 
         for name in type(self).model_fields:
             _validate(self, name)
+        return self
+
+    @model_validator(mode="after")
+    def log_settings(self) -> Self:
+        """Validator that logs the settings in text form."""
+        logger.info(
+            click.style(
+                f"{type(self).__name__}\n{self.text()}",
+                fg="bright_cyan",
+            )
+        )
         return self